+++ lib/src/models/request/a2a_eft_cancel_request.dart
@@ -16,16 +16,17 @@ sealed class PPEftCancelRequestModel with _$PPEftCancelRequestModel {
       _$PPEftCancelRequestModelFromJson(json);
 
   factory PPEftCancelRequestModel.toRequest({
-    
     required String clientToken,
     required String transactionId,
     required double totalAmount,
+    PPDeviceTypeEnum deviceType = PPDeviceTypeEnum.POS,
   }) => PPEftCancelRequestModel(
     data: PPEftCancelRequestData(totalAmount: totalAmount),
     header: PPGeneralRequestHeader(
       transactionType: PPTransactionType.POST_EFTPOS_CANCEL,
       clientToken: clientToken,
       transactionId: transactionId,
+      deviceType: deviceType,
     ),
   );
 }