+++ lib/src/models/request/a2a_eft_payment_request.dart
@@ -22,6 +22,7 @@ sealed class PPEftPaymentRequestModel with _$PPEftPaymentRequestModel {
     required PPPaymentMethod paymentMethod,
     required String transactionId,
     required int taxRate,
+    PPDeviceTypeEnum deviceType = PPDeviceTypeEnum.POS,
     int? installment,
   }) => PPEftPaymentRequestModel(
     data: PPEftPaymentRequestData(
@@ -35,6 +36,7 @@ sealed class PPEftPaymentRequestModel with _$PPEftPaymentRequestModel {
     header: PPGeneralRequestHeader(
       transactionType: PPTransactionType.POST_EFTPOS,
       clientToken: clientToken,
+      deviceType: deviceType
     ),
   );
 }