+++ lib/src/models/request/a2a_request_header.g.dart
@@ -14,6 +14,9 @@ _PPGeneralRequestHeader _$PPGeneralRequestHeaderFromJson(
     json['transaction_type'],
   ),
   clientToken: json['client_token'] as String,
+  deviceType:
+      $enumDecodeNullable(_$PPDeviceTypeEnumEnumMap, json['device_type']) ??
+      PPDeviceTypeEnum.POS,
   orderCode: json['order_code'] as String?,
   transactionId: json['transaction_id'] as String?,
 );
@@ -23,6 +26,7 @@ Map<String, dynamic> _$PPGeneralRequestHeaderToJson(
 ) => <String, dynamic>{
   'transaction_type': _$PPTransactionTypeEnumMap[instance.transactionType]!,
   'client_token': instance.clientToken,
+  'device_type': _$PPDeviceTypeEnumEnumMap[instance.deviceType]!,
   'order_code': instance.orderCode,
   'transaction_id': instance.transactionId,
 };
@@ -35,4 +39,9 @@ const _$PPTransactionTypeEnumMap = {
   PPTransactionType.ORDER_PAYMENT: 'ORDER_PAYMENT',
   PPTransactionType.EOD: 'EOD',
   PPTransactionType.PARAMETERS: 'PARAMETERS',
+};
+
+const _$PPDeviceTypeEnumEnumMap = {
+  PPDeviceTypeEnum.POS: 'POS',
+  PPDeviceTypeEnum.KIOSK: 'KIOSK',
 };