+++ lib/src/models/request/a2a_start_payment_request.dart
@@ -28,6 +28,7 @@ sealed class PPStartPaymentRequestModel with _$PPStartPaymentRequestModel {
     PPPartialPaymentType? partialType,
     List<PPProduct> products = const [],
     PPBillingInfo? billingInformation,
+    PPDeviceTypeEnum deviceType = PPDeviceTypeEnum.POS,
   }) => PPStartPaymentRequestModel(
     data: PPStartPaymentRequestData(
       paymentType: paymentType,
@@ -43,6 +44,7 @@ sealed class PPStartPaymentRequestModel with _$PPStartPaymentRequestModel {
       transactionType: PPTransactionType.POST_PAYMENT_START,
       clientToken: clientToken,
       orderCode: orderCode,
+      deviceType: deviceType,
     ),
   );
 }